# cert-api2/public/swagger.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Taichung School Certificate Project
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
  - url: http://localhost:3000/api/test
  - url: http://192.168.56.101:3000/api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/oid-tcsso-authurl:
    post:
      summary: 台中市教育局校園登入 OPENID authorization_endpoint
      tags:
        - authorization
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                response_type: string
                azp: string
                redirect_uri: string
                scope: string
                state: number
              example:
                response_type: code
                scope: openid+email
      responses:
        '200':
          content:
            application/json:
  /auth/jwt-token:
    post:
      summary: get jwt token
      tags:
        - authorization
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: number
                username: string
                roles: array
              example:
                id: 1
                username: bob
                roles: ['USER']
      responses:
        '200':
          content:
            application/json:
  /auth/jwt-authorized:
    post:
      summary: jwt token authorized
      tags:
        - authorization
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: number
              example:
                id: 1
      responses:
        '200':
          content:
            application/json:
...
  /auth/oid-callback:
    post:
      summary: 進行中 OpenID Connect 1.0 # 3.1.2 (http://openid.net/specs/openid-connect-core-1_0.html#AuthorizationEndpoint)
      tags:
        - authorization
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...